# openapi.yaml
openapi: 3.1.0
info:
  title: Sportsdeck API
  version: 1.0.0
  description: API documentation for Sportsdeck, including endpoints for authentication, users, matches, forums, polls, and moderation.
servers:
  - url: /api
    description: Base API endpoint
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /threads/{threadId}:
    get:
      summary: Get thread details
      description: View a specific thread along with its main post and replies.
      security: []
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thread object
        '404':
          description: Thread not found or hidden
        '500':
          description: Fetch failed

    put:
      summary: Update thread
      description: Edit a thread's title and/or main content. A version history is saved on every update. Author or Admin only.
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
            example:
              content: "Updated content..."
      responses:
        '200':
          description: Updated thread object
        '400':
          description: Nothing to update
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not author)
        '404':
          description: Thread not found or hidden
        '500':
          description: Update failed

    delete:
      summary: Delete thread
      description: Soft-delete a thread. Author or Admin only.
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              example:
                success: true
                message: "Thread deleted"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Thread not found
        '500':
          description: Deletion failed
...
  /threads/{threadId}/replies:
    get:
      summary: Get thread replies
      description: Fetch replies associated with this specific thread.
      security: []
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Array of post reply objects
        '500':
          description: Fetch failed

    post:
      summary: Reply to thread
      description: Post a reply to an active thread. The reply is asynchronously scanned for toxicity via AI.
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
            example:
              content: "I agree with this."
      responses:
        '201':
          description: Created reply post object
        '400':
          description: Missing content
        '401':
          description: Unauthorized
        '403':
          description: Thread not found or is locked/hidden
        '500':
          description: Creation failed
...
  /threads/{threadId}/translation:
    get:
      summary: Translate a thread
      description: Translate the main content of a specific thread into the target language using AI.
      security: []
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: string
        - in: query
          name: lang
          schema:
            type: string
            default: "English"
      responses:
        '200':
          description: Translation successful
          content:
            application/json:
              example:
                translation: "Translated string"
        '403':
          description: Cannot translate hidden content
        '404':
          description: Thread not found
        '500':
          description: Translation failed
...
  /reports:
    post:
      summary: Submit a report
      description: Submit a new report about a thread or post.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - resourceId
                - resourceType
                - reason
              properties:
                resourceId:
                  type: string
                resourceType:
                  type: string
                  enum: [THREAD, POST]
                reason:
                  type: string
            example:
              resourceId: "thread123"
              resourceType: "THREAD"
              reason: "Inappropriate content"
      responses:
        '200':
          description: Report submitted
        '400':
          description: Missing fields or invalid type
        '401':
          description: Unauthorized
        '404':
          description: Target resource not found
        '500':
          description: Submission failed

    get:
      summary: Get reports
      description: Return paginated aggregated report buckets. Admin only.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated reports
          content:
            application/json:
              example:
                data: []
                pagination:
                  page: 1
                  limit: 20
                  totalItems: 0
                  totalPages: 1
        '401':
          description: Unauthorized
        '500':
          description: Fetch failed
...